requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
idna==3.10
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pdf2image==1.17.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyMuPDF==1.26.4
pypng==0.20220715.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
qrcode==7.4.2
reportlab==4.4.4
requests==2.32.5
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0

# Production dependencies
gunicorn==22.0.0
whitenoise==6.6.0
